refactor(coin-zcash): drop the CoinModuleApi surface
The api needed an address-indexed data source the AccountBridge never exercises,
and one that cannot answer for shielded value at all, since notes are encrypted
to a viewing key an address argument does not carry. Migrating to it becomes a
separate task, so that switching away from the coin-bitcoin adapter changes
nothing but where the code lives.

Removes src/api and the logic barrel, plus everything that existed only for
them: listOperations, lastBlock, getNextSequence, the estimateFees wrapper over
coin-selection, the explorer client, the address balance reader and the
transaction classifier. logic/ now holds only what bridge/ consumes.
