+++ libs/coin-modules/coin-zcash/src/logic/history/transparentTx.unit.test.ts
@@ -1,11 +1,5 @@
 import type { TX } from "@ledgerhq/wallet-btc/index";
-import {
-  classifyTransparentTx,
-  explorerFee,
-  spentOutpoints,
-  sumValues,
-  txDate,
-} from "./transparentTx";
+import { explorerFee, spentOutpoints, txDate } from "./transparentTx";
 
 const OWN = "t1OwnAddress";
 const OTHER = "t1OtherAddress";
@@ -33,16 +27,6 @@ const input = (address: string | undefined, value: string, outputIndex = 0) =>
     output_index: outputIndex,
     sequence: 0xfffffffe,
   }) as TX["inputs"][number];
-
-const output = (address: string, value: string, outputIndex = 0) =>
-  ({
-    address,
-    value,
-    output_hash: "cc".repeat(32),
-    output_index: outputIndex,
-    block_height: 3_000_000,
-    rbf: false,
-  }) as TX["outputs"][number];
 
 describe("spentOutpoints", () => {
   it("lists the outpoints spent, in `${txid}-${index}` form", () => {
@@ -76,69 +60,5 @@ describe("txDate", () => {
 
   it("falls back to the reception time when unconfirmed", () => {
     expect(txDate(tx({ block: null })).toISOString()).toBe("2026-01-02T03:04:05.000Z");
-  });
-});
-
-describe("sumValues", () => {
-  it("sums as decimal, not as float", () => {
-    expect(sumValues([{ value: "9007199254740993" }, { value: "1" }]).toString()).toBe(
-      "9007199254740994",
-    );
-  });
-
-  it("sums nothing to zero", () => {
-    expect(sumValues([]).toString()).toBe("0");
-  });
-});
-
-describe("classifyTransparentTx", () => {
-  const isOwn = (address: string) => address === OWN;
-
-  it("splits what the owner spent from what came back", () => {
-    const participation = classifyTransparentTx(
-      tx({
-        inputs: [input(OWN, "5000"), input(OTHER, "1000", 1)],
-        outputs: [output(OTHER, "3000"), output(OWN, "1800", 1)],
-      }),
-      isOwn,
-    );
-
-    expect(participation.spent.toString()).toBe("5000");
-    expect(participation.returned.toString()).toBe("1800");
-    expect(participation.ownInputs).toHaveLength(1);
-    expect(participation.ownOutputs).toHaveLength(1);
-  });
-
-  it("reports senders and recipients without duplicates, in transaction order", () => {
-    const participation = classifyTransparentTx(
-      tx({
-        inputs: [input(OTHER, "1000"), input(OTHER, "2000", 1)],
-        outputs: [output(OWN, "1500"), output(OWN, "1400", 1)],
-      }),
-      isOwn,
-    );
-
-    expect(participation.senders).toEqual([OTHER]);
-    expect(participation.recipients).toEqual([OWN]);
-  });
-
-  it("ignores an input the explorer could not attribute", () => {
-    const participation = classifyTransparentTx(
-      tx({ inputs: [input(undefined, "5000")], outputs: [output(OWN, "4000")] }),
-      isOwn,
-    );
-
-    expect(participation.spent.toString()).toBe("0");
-    expect(participation.senders).toEqual([]);
-  });
-
-  it("drops an output the explorer reports as unknown", () => {
-    const participation = classifyTransparentTx(
-      tx({ outputs: [output("unknown_2", "4000"), output(OWN, "1000", 1)] }),
-      isOwn,
-    );
-
-    expect(participation.recipients).toEqual([OWN]);
-    expect(participation.returned.toString()).toBe("1000");
   });
 });
